refactor: Move utils to test fixtures
