+++ nvi-commons/src/test/java/no/sikt/nva/nvi/common/db/PeriodRepositoryTest.java
@@ -1,6 +1,6 @@
 package no.sikt.nva.nvi.common.db;
 
-import static no.sikt.nva.nvi.test.TestUtils.randomNviPeriodBuilder;
+import static no.sikt.nva.nvi.common.db.DbNviPeriodFixtures.randomNviPeriodBuilder;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
@@ -24,8 +24,8 @@ class PeriodRepositoryTest extends LocalDynamoTest {
   void shouldUpdateExistingRecordWhenSavedSecondTime() {
 
     var year = TestUtils.randomYear();
-    var user1 = TestUtils.randomUsername();
-    var user2 = TestUtils.randomUsername();
+    var user1 = UsernameFixtures.randomUsername();
+    var user2 = UsernameFixtures.randomUsername();
 
     var nviPeriod1 = randomNviPeriodBuilder().publishingYear(year).modifiedBy(user1).build();
     var nviPeriod2 = randomNviPeriodBuilder().publishingYear(year).modifiedBy(user2).build();
